Module trie::set

An ordered set based on a trie.

Structs

Difference
An iterator producing elements in the set difference (in-order).

Intersection
An iterator producing elements in the set intersection (in-order).

Iter
A forward iterator over a set.

Range
A bounded forward iterator over a set.

Set
A set implemented as a radix trie.

SymmetricDifference
An iterator producing elements in the set symmetric difference (in-order).

Union
An iterator producing elements in the set union (in-order).