Struct trie::TrieSet

pub struct TrieSet { /* private fields */ }

A set base on a trie.

Implementations

impl TrieSet

pub fn new() -> Self

Makes a new empty TrieMap.

Does not allocate anything on its own.

Examples

Basic usage:

use trie::TrieSet;
use std::iter::once;

let mut set = TrieSet::new();

// entries can now be inserted into the empty set
set.insert(once(1));

pub fn contains(&self, value: impl IntoIterator<Item = usize>) -> bool

Returns true if the set contains a value.

Examples

use trie::TrieSet;
use std::iter::once;

let mut set = TrieSet::new();
set.insert(once(1));
set.insert(once(2));
set.insert(once(3));

// let set: TrieSet<_> = [1, 2, 3].iter().cloned().collect(); TODO: unimplemented
assert_eq!(set.contains(once(1)), true);
assert_eq!(set.contains(once(4)), false);

pub fn insert(&mut self, iter: impl IntoIterator<Item = usize>) -> bool

Adds a value to the set.

If the set did not have this value present, true is returned.

If the set did have this value present, false is returned, and the entry is not updated.

Examples

use std::iter::once;
use trie::TrieSet;

let mut set = TrieSet::new();

assert_eq!(set.insert(once(2)), true);
assert_eq!(set.insert(once(2)), false);
// assert_eq!(set.len(), 1); not implemented

pub fn remove(&mut self, value: impl IntoIterator<Item = usize>) -> bool

Removes a value from the set. Returns whether the value was present in the set.

Examples

use std::iter::once;
use trie::TrieSet;

let mut set = TrieSet::new();

set.insert(once(2));
assert_eq!(set.remove(once(2)), true);
assert_eq!(set.remove(once(2)), false);

pub fn for_each_prefix( &self, value: impl IntoIterator<Item = usize>, visit: impl FnMut(&TrieMap<()>) )

Visits all the “existing” nodes corresponding to the preficies of the value.

Examples

Basic usage:

use trie::TrieSet;
use std::iter::once;

let mut set = TrieSet::new();
set.insert(vec![1]);
set.insert(vec![1, 1, 1]);

// Corresponding an existing value.
let mut expected = [false, true, false, true].iter();
set.for_each_prefix(vec![1, 1, 1].into_iter(), |trie| {
    let expected = *expected.next().unwrap();
    assert_eq!(trie.get(None.into_iter()).is_some(), expected);
});

// No, but falls within the trie.
let mut expected = [false, true, false].iter();
set.for_each_prefix(vec![1, 1].into_iter(), |trie| {
    let expected = *expected.next().unwrap();
    assert_eq!(trie.get(None.into_iter()).is_some(), expected);
});

// Runs off thte trie.
let mut expected = [false, true, false, true].iter();
set.for_each_prefix(vec![1, 1, 1, 1].into_iter(), |trie| {
    let expected = *expected.next().unwrap();
    assert_eq!(trie.get(None.into_iter()).is_some(), expected);
});

pub fn for_each(&self, visit: impl FnMut(&[usize]))

Visits all the values of a values in the trie, in lexicographic order.

Examples

Basic usage:

use trie::TrieSet;
use std::iter::once;

let mut set = TrieSet::new();
set.insert(vec![1, 2, 2]);
set.insert(vec![2]);
set.insert(vec![1]);
set.insert(vec![1, 1, 1]);

let mut expected = vec![
    vec![1],
    vec![1, 1, 1],
    vec![1, 2, 2],
    vec![2],
].into_iter();
set.for_each(|k| {
    let ek = expected.next().unwrap();
    assert_eq!(k, ek.as_slice());
});